OPR后台逻辑，删除冗余class

+++ pmc_report/pmc_report_manager/src/main/java/gean/pmc_report_manager/modules/report/vo/AreaOprVo.java
@@ -9,7 +9,7 @@ public class AreaOprVo {
 
 	private String area;
 	private Integer actual;
-	private Float shiftPlan;
+	private Integer shiftPlan;
 	private String zone;
 	private Integer variation;
 	private Float productionOpr;
